Active tutors show on home page
Edited base.html to be more organized while maintaining all original functionality. Edited home.html to show a list of all active tutors. It is not yet possible to interact with this list.
Edited views to pass in a list of all tutors to home.html which will then parse the list and print out all active tutors.